ai_worker/requirements.txt
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
llama-cpp-python>=0.2.11
redis>=5.0.1
pydantic-settings>=2.2.1
psycopg2-binary>=2.9.9
numpy==1.26.4
requests>=2.31.0
bleach>=6.1.0
